fx: Fix config load order must be reversed
Overload with outside-in loading order is actually wrong because it
would ignores the env var in the current environment (.env > ENV)

Just blindly changing it to Load is still wrong because outside envs
then cannot be overridden from the inside. (.env > child/.env)

The correct way is to use `Load` with inside-out order so priority goes
like this:

  Env > child/.env.local > child/.env > .env.local > .env > default

The file `correct-output.txt` and a testcase has also been added.