fx: Fix connection leak in data middleware
Happens when we have multiple `Fragment` along with data middleware
being added during fragment composition, resulting in multiple
`dataContext` being created.

This fixes so that multiple call to middleare composition routine would
not create multiple `dataContext`, only if we call the `AddDataContext`
repeatedly.

We have to review the fragment code again, to make sure that didn't
happen or that the contract is clear.